refactor: split helpers into proxy and event
