+++ contracts/utils/proxy.ts
@@ -1,43 +1,7 @@
 import {ProxyFactory__factory} from '../typechain';
 import {ProxyCreatedEvent} from '../typechain/src/utils/ProxyFactory';
-import {ContractFactory, ContractTransaction} from 'ethers';
-import {
-  Interface,
-  LogDescription,
-  defaultAbiCoder,
-  keccak256,
-} from 'ethers/lib/utils';
-import {ethers} from 'hardhat';
-
-export function toBytes(string: string) {
-  return ethers.utils.formatBytes32String(string);
-}
-
-export function hashHelpers(helpers: string[]) {
-  return keccak256(defaultAbiCoder.encode(['address[]'], [helpers]));
-}
-
-export async function findEvent<T>(tx: ContractTransaction, eventName: string) {
-  const receipt = await tx.wait();
-
-  const event = (receipt.events || []).find(event => event.event === eventName);
-
-  return event as T | undefined;
-}
-
-export async function findEventTopicLog<T>(
-  tx: ContractTransaction,
-  iface: Interface,
-  eventName: string
-): Promise<LogDescription & (T | LogDescription)> {
-  const receipt = await tx.wait();
-  const topic = iface.getEventTopic(eventName);
-  const log = receipt.logs.find(x => x.topics[0] === topic);
-  if (!log) {
-    throw new Error(`No logs found for the topic of event "${eventName}".`);
-  }
-  return iface.parseLog(log) as LogDescription & (T | LogDescription);
-}
+import {findEvent} from './events';
+import {ContractFactory} from 'ethers';
 
 export async function deployUUPSProxy<T>(
   contractFactory: ContractFactory,
@@ -90,5 +54,5 @@ export async function deployMinimalClone<T>(
     throw new Error('Failed to get the event');
   }
 
-  return contractFactory.attach(event.args.clone) as unknown as T;
+  return contractFactory.attach(event.args.proxy) as unknown as T;
 }
